requirements.txt
mcstatus
requests
colorama
fake-useragent
mcrcon
dnspython
toml
pyperclip
PySocks
keyboard
mysql-connector-python
bcrypt
